feat: make sessionauth throw if onFailureRedirectionURL returns a relative path
